refactor: cut out RouteInfo from route.rs
